+++ src/containers/StockBord/StockBord.module.css
@@ -24,26 +24,6 @@ li {
   }
 }
 
-.content {
-  width: 150px;
-  height: 50px;
-  background-color: rgb(56, 56, 56);
-  padding-top: 15px;
-  color: aliceblue;
-}
-
 .slide-paused:hover .slideshow {
   animation-play-state: paused;
-}
-
-.red {
-  background-color: rgb(235, 0, 0);
-  border-radius: 4px;
-  padding: 0px 10px;
-}
-
-.green {
-  background-color: rgb(0, 187, 0);
-  border-radius: 4px;
-  padding: 0px 10px;
 }
